services/telemetry-processor/requirements.txt
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0